hopeless to use a newer version of tramp
Either I am unable to load it cleanly, because the system tramp still
is loaded, at least parts of it. Using el-get leads to the weirdes
errors. Using a manually built git version doesn't work either. It
loads but then I can't ssh anywhere.

So, I'll stick to the default and wait for the ubuntu emacs update.