Stupid me!
A couple of  years ago I added  this hydra and forgone  my beloved M-o
command to cycle  through all windows, because I  thought this is
better. However, whenever I have more than 2 windows and need to
switch, this hydra appears and I hesitate, because the workflow just
doesn't fit to my muscle memory for whatever reasons.

Now I just added M-o to the list of subcommands which cycles through
the windows! Just as in the earlier days! I am so happy with it and in
the same time embarrased of myself that I didn't think earlier
about this solution!